fix: remove insecure part-tools stub and address edge cases
- Deleted `mastra/tools/part-tools.ts` which contained an insecure stub implementation with `STUB_USER_ID`.
- Renamed `mastra/tools/part-tools.mastra.ts` to `mastra/tools/part-tools.ts` to serve as the canonical implementation.
- Updated `mastra/agents/ifs-agent.ts` to import from the correct location.
- Updated `config/dev.ts` to provide a clear error message ("User ID is required") in production when `userId` is missing, instead of suggesting development variables.
- Removed redundant `evidence.length < 3` check in `lib/data/schema/parts-agent.ts` as it is enforced by Zod schema.
- Verified changes with `npm run typecheck` and `npm test:unit`.
